+++ src/client/request/graphql/adapters/credit-card-tokenization.js
@@ -22,6 +22,13 @@ var BIN_DATA_MAP = {
   UNKNOWN: 'Unknown'
 };
 
+var AUTHENTICATION_INSIGHT_MAP = {
+  UNKNOWN: 'unknown',
+  PSDTWO: 'psd2',
+  UNREGULATED: 'unregulated',
+  UNAVAILABLE: 'unavailable'
+};
+
 function creditCardTokenizationResponseAdapter(responseBody) {
   var adaptedResponse;
 
@@ -39,7 +46,7 @@ function adaptTokenizeCreditCardResponseBody(body) {
   var creditCard = data.creditCard;
   var lastTwo = creditCard.last4 ? creditCard.last4.substr(2, 4) : '';
   var binData = creditCard.binData;
-  var response;
+  var response, regulationEnvironment;
 
   if (binData) {
     ['commercial', 'debit', 'durbinRegulated', 'healthcare', 'payroll', 'prepaid'].forEach(function (key) {
@@ -77,8 +84,9 @@ function adaptTokenizeCreditCardResponseBody(body) {
   };
 
   if (data.authenticationInsight) {
+    regulationEnvironment = data.authenticationInsight.customerAuthenticationRegulationEnvironment;
     response.creditCards[0].authenticationInsight = {
-      regulationEnvironment: data.authenticationInsight.customerAuthenticationRegulationEnvironment
+      regulationEnvironment: AUTHENTICATION_INSIGHT_MAP[regulationEnvironment] || AUTHENTICATION_INSIGHT_MAP.UNKNOWN
     };
   }