+++ src/hosted-fields/shared/browser-detection.js
@@ -4,6 +4,9 @@ module.exports = {
   isIE: require('@braintree/browser-detection/is-ie'),
   isEdge: require('@braintree/browser-detection/is-edge'),
   isIe9: require('@braintree/browser-detection/is-ie9'),
+  isIe10: require('@braintree/browser-detection/is-ie10'),
+  isAndroid: require('@braintree/browser-detection/is-android'),
+  isChromeOS: require('@braintree/browser-detection/is-chrome-os'),
   isIos: require('@braintree/browser-detection/is-ios'),
   isIosWebview: require('@braintree/browser-detection/is-ios-webview')
 };